backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-environ==0.11.2
psycopg2-binary==2.9.7
Pillow==10.0.1
python-dotenv==1.0.0
setuptools==68.0.0